Unpin xarray version [all tests ci] (#981)
* Unpin xarray version

* unpin xarray from yaml

* Change mamba env to pull from reqs to sync deps

* Fix path to requirements.txt

* Fix vend_cal_params_power function and expected output type

* Fix linting for cal_params

* Retrict the variables used during interpolation

* Pin pandas to <2

* Add dtype sanitizer before figuring out encodings

* Use sanitize dtype before writing to zarr

* Set psutil to be able to install latest

+++ echopype/tests/calibrate/test_cal_params.py
@@ -627,7 +627,7 @@ def test_get_cal_params_EK60(beam_EK, vend_EK, freq_center, user_dict, out_dict)
                 dims=["ping_time", "channel"],
                 coords={"ping_time": [1, 2, 3, 4], "channel": ["chA", "chB"]},
                 name="sa_correction",
-            ),
+            ).astype(np.float64),
         ),
         # with NaN entry in transmit_duration_nominal
         (
